requirements.txt
numpy >= 1.17
scipy >= 1.5
numba == 0.48
tensorflow < 2.0
cupy == 9.2
netifaces >= 0.11.0
paho-mqtt >= 1.6.1
numpyencoder >= 0.3.0
Pillow >= 8.4.0
python-engineio >= 3.14.2
python-socketio >= 4.3.1